requirements.txt
pandas
numpy
prophet
scikit-learn
streamlit
plotly
mysql-connector-python
python-dotenv

